Move logger code to module __init__
