+++ plugin.py
@@ -33,44 +33,7 @@ import os
 # Create the logger for this QgisNetworkLogger plugin
 import logging
 from . import LOGGER_NAME
-from qgis.core import (
-    Qgis,
-    QgsMessageLog
-)
-class QgisLogHandler(logging.StreamHandler):
-    '''
-    Some magic to make it possible to use code like:
-
-    import logging
-    from . import LOGGER_NAME
-    log = logging.getLogger(LOGGER_NAME)
-
-    in all this plugin code, and it will show up in the QgsMessageLog
-
-    '''
-    def __init__(self, topic):
-        logging.StreamHandler.__init__(self)
-        # topic is used both as logger id and for tab
-        self.topic = topic
-
-    def emit(self, record):
-        msg = self.format(record)
-        # Below makes sure that logging of 'self' will show the repr of the object
-        # Without this it will not be shown because it is something like
-        # <qgisnetworklogger.plugin.QgisNetworkLogger object at 0x7f580dac6b38>
-        # which looks like an html element so is not shown in the html panel
-        msg=msg.replace('<', '&lt;').replace('>', '&gt;')
-        QgsMessageLog.logMessage('{}'.format(msg), self.topic, Qgis.Info)
-
-
 log = logging.getLogger(LOGGER_NAME)
-# checking below is needed, else we add this handler every time the plugin
-# is reloaded (during development), then the msg is emitted several times
-if not log.hasHandlers():
-    log.addHandler(QgisLogHandler(LOGGER_NAME))
-
-# set logging level (NOTSET = no, else: DEBUG or INFO)
-log.setLevel(logging.DEBUG)
 
 class QgisNetworkLogger:
     '''
